% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climato.R
\name{compute_anomalies}
\alias{compute_anomalies}
\title{Computing anomalies.}
\usage{
compute_anomalies(
  data,
  clim_data,
  years_range = NULL,
  anomaly_type = c("difference", "standardized")
)
}
\arguments{
\item{data}{object of class \code{scda_data}.}

\item{clim_data}{object of class \code{scda_climato}, the climatology data computed from the function \code{\link[simpleCDA]{compute_climatology}}.}

\item{years_range}{an integer vector of length 2, the start and end year of the period the anomalies will be computed.
Default \code{NULL}, using the whole period available in the dataset.}

\item{anomaly_type}{character, it specifies the type of anomalies to be calculated 
and must be one of \code{"difference"} and \code{"standardized"}.}
}
\value{
This returns an object of class \code{scda_anomalies} and \code{scda_data}. The object has the following elements:
\itemize{
\item \code{$lon}: a numeric vector of the longitude with length \code{N}.
\item \code{$lat}: a numeric vector of the latitude with length \code{M}.
\item \code{$date}: a vector of class \code{Date} with length \code{L}.
\item \code{$data}: a list of matrices with dimension \code{N x M}, the length of the list is \code{L}.
}
And other information from the input data.
}
\description{
Compute anomalies for a specific month or season.
}
\examples{
\dontrun{
# Standardized anomaly of the SST for the season DJF for the period 1981-2020.

# read the data
path_dir_data = '/home/data/ERSSTv5_seasonal'
nc_file_fromat = 'ersstv5_\%s-\%s_\%s-\%s.nc'
grd_data = get_season_region_netcdf_data('12-02', path_dir_data, nc_file_fromat,
                                         years_range = c(1981, 2020))

# compute climatology for the period 1991 to 2020
clim_data = compute_climatology(grd_data, base_period = c(1991, 2020))

# compute the standardized anomaly
anom_data = compute_anomalies(grd_data, clim_data, 
                              years_range = c(1981, 2020),
                              anomaly_type = 'standardized')
}

}
