% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climato.R
\name{compute_climatology}
\alias{compute_climatology}
\title{Computing climatology.}
\usage{
compute_climatology(data, base_period = c(1991, 2020))
}
\arguments{
\item{data}{object of class \code{scda_data}.}

\item{base_period}{an integer vector of length 2, giving the start and end year of the base period to compute the climatology.}
}
\value{
This returns an object of class \code{scda_climato}. The object has the following elements:
\itemize{
\item \code{$lon}: a numeric vector of the longitude with length \code{N}.
\item \code{$lat}: a numeric vector of the latitude with length \code{M}.
\item \code{$base_period}: the base period used to compute the climatology.
\item \code{$data}: a list of 2 matrices of the \strong{mean} and \strong{standard deviation}, with dimension \code{N x M}.
}
And other information from the input data.
}
\description{
Compute climatology for a specific month or season.
}
