% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigens.R
\name{get_EOFs_spatialPatterns}
\alias{get_EOFs_spatialPatterns}
\title{Extracting the spatial patterns (EOFs).}
\usage{
get_EOFs_spatialPatterns(eigen_data, k)
}
\arguments{
\item{eigen_data}{object of class \code{scda_eigen}.}

\item{k}{integer, the \code{nth} EOFs to be extracted.}
}
\value{
This returns an object of class \code{scda_eigen_eofs}.
The object has the following elements:
\itemize{
\item \code{$lon}: a numeric vector of the longitude with length \code{N}.
\item \code{$lat}: a numeric vector of the latitude with length \code{M}.
\item \code{$eof}: a matrix of the \code{nth} spatial patterns (EOFs) with dimension \code{N x M}.
}
}
\description{
Extract a specific spatial patterns (EOFs).
}
\examples{
\dontrun{
## grd_data: object of class "scda_data"

## perform eigenanalysis
eigen_data = eigen_analysis(grd_data)

## get the first EOFs
eof1 = get_EOFs_spatialPatterns(eigen_data, 1)
}

}
