% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigens.R
\name{get_PCs_timeSeries}
\alias{get_PCs_timeSeries}
\title{Extracting the principal component (PCs).}
\usage{
get_PCs_timeSeries(eigen_data, k)
}
\arguments{
\item{eigen_data}{object of class \code{scda_eigen}.}

\item{k}{integer, the \code{nth} PCs to be extracted.}
}
\value{
This returns an object of class \code{scda_eigen_pcs}.
The object has the following elements:
\itemize{
\item \code{$date}: a vector of class \code{Date}.
\item \code{$pc}: a numeric vector of the \code{nth} time series (PCs).
}
}
\description{
Extract a specific time series (PCs).
}
\examples{
\dontrun{
## grd_data: object of class "scda_data"

## perform eigenanalysis
eigen_data = eigen_analysis(grd_data)

## get the second PCs
pc2 = get_PCs_timeSeries(eigen_data, 2)
}

}
