% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigens.R
\name{plot_EOFs_spatialPatterns}
\alias{plot_EOFs_spatialPatterns}
\title{Display the spatial patterns (EOFs).}
\usage{
plot_EOFs_spatialPatterns(
  eigen_data,
  k,
  zlim = c(-0.05, 0.05),
  col_palette = color_Brewer_Fun,
  ...,
  add_map = TRUE,
  map_options = list(col = "gray95", fill = TRUE, lwd = 0.2, border = "black")
)
}
\arguments{
\item{eigen_data}{object of class \code{scda_eigen}.}

\item{k}{integer, the \code{nth} EOFs to be displayed.}

\item{zlim}{a numeric vector of length 2, giving the minimum and maximum limit of data to display.
The values under or above these limits are masked with the corresponding colors assigned to these limits.}

\item{col_palette}{the color palette function to be used. User defined color palette or a predefined color palettes from the package \pkg{grDevices} or other packages.}

\item{...}{other arguments that may be passed on to the function \code{\link[fields]{image.plot}} of the package \pkg{fields}.}

\item{add_map}{logical, default \code{TRUE} add world map to the plot.}

\item{map_options}{a named list, if \code{add_map = TRUE}, a list of options to be passed on the function \code{\link[maps]{map}} of the package \pkg{maps}.}
}
\description{
Plot a specific spatial patterns (EOFs) on map.
}
\examples{
\dontrun{
path_dir_data = '/home/data/ERSSTv5'
nc_file_fromat = 'ersst.v5.\%s\%s.nc'
years_range = c(1981, 2020)
grd_data = get_month_region_netcdf_data(12, path_dir_data, nc_file_fromat, years_range)
eigen_data = eigen_analysis(grd_data)
plot_EOFs_spatialPatterns(eigen_data, 2)
}
 
}
